Changed Async storage of registration values to only occur at the end
